Refactor structure, fix setup/docs, and squash latent bugs
Structure / professionalism:
- Rename misleading entrypoints: scripts/infer/test.py -> run.py,
  scripts/pipeline/test_llm.py -> phishvlm.py (class TestVLM -> PhishVLM),
  scripts/phishintention/configs.py -> model_config.py (removes the
  configs.py vs configs/ package-name clash).
- Add explicit __init__.py files so the `scripts` package no longer
  depends on implicit namespace-package resolution.
- run.py: replace wildcard imports with explicit imports, drop the forced
  CUDA_LAUNCH_BLOCKING debug flag and dead openai.api_key/proxy lines.

Bug fixes:
- .gitignore: `./datasets/*` and `./scripts/.../models/*` patterns were
  invalid (leading `./`), so API-key files were NOT ignored. Replace with
  valid, scoped patterns that protect secrets while keeping the demo data.
- setup.sh: define the previously-undefined error_exit helper.
- web_utils.query2image: guard against items lacking an `image` field.
- stn_head.py: `is 'none'` -> `== 'none'` (identity vs equality).
- run.py: strip the URL read from info.txt; default --folder to the
  bundled demo; clearer FileNotFoundError for a missing OpenAI key.

Docs / deps:
- README: corrected run command (`python -m scripts.infer.run`), added the
  missing Chrome download step, accurate repo tree, config/troubleshooting
  sections, and a decision logic that matches the implementation.
- requirements.txt: add missing PyYAML, pin openai>=1.13 (v1 client),
  group and document the torch/detectron2 out-of-band install.
